+++ src/utils/check-auth.js
